requirements.txt
requests==2.18.4
numpy==1.14.3
tqdm==4.23.4
pillow==5.1.0
scikit-learn==0.19.1
keras==2.2.0
tensorflow==1.8.0